experiment(#2): pool_heap files are pool; tracker_create is the name
One afw_pool.c / afw_pool_internal.h. ST heap is create_xctx_p.
afw_pool_tracker_create replaces heap_tracker_create. Call sites
updated. Test dir still pool_heap.
